pyproject.toml
[tool.poetry]
name = "etl-duckdb"
version = "0.1.0"
description = ""
authors = ["wesllanSilva <“wesllan90@gmail.com”>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
gdown = "^5.2.0"
streamlit = "^1.41.0"
duckdb = "^1.1.3"
sqlalchemy = "^2.0.36"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.10"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
